% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{chip_galonska}
\alias{chip_galonska}
\title{Another example ChIP data set that can be used 
with \code{\link{importDataFromMatrices}} from rom Galonska et. al., 2015.}
\format{a matrix}
\usage{
data(chip_galonska)
}
\value{
a matrix
}
\description{
The containes the H3Kme3 data in the serum and 24h_2i conditions 
from Galonska et. al., 2015
as well as the whole cell extract data, which is treated as input for all
four samples.
The data were downloaded from the SRA at the european
nucleotide archive (ENA, accession PRJNA242892).
The reads were aligned ot the mm9 reference genome using bowtie2
(Langmead and Salzberg, 2012) with
default options. Then, filtering of unmapped, low mapping quality (< 10), 
duplicated and
multi-mapping reads was performed with Picard tools.
The fragment length was inferred using
cross correlation plots from SPP (Kharchenko, et. al., 2008).
}
\references{
Galonska, Christina, Michael J. Ziller, Rahul Karnik, and Alexander 
Meissner. 2015. "Ground State Conditions Induce Rapid Reorganization of 
Core Pluripotency Factor Binding Before Global Epigenetic Reprogramming." 
Cell Stem Cell 17 (4). Elsevier BV: 462-70.
\url{http://dx.doi.org/10.1016/j.stem.2015.07.005}.
Kharchenko, Peter V, Michael Y Tolstorukov, and Peter J Park. 2008. 
"Design and Analysis of ChIP-Seq Experiments for DNA-Binding Proteins."
Nat Biotechnol 26 (12). Nature Publishing Group: 1351-9. 
\url{http://dx.doi.org/10.1038/nbt.1508}.
Langmead, Ben, and Steven L Salzberg. 2012. 
"Fast Gapped-Read Alignment with Bowtie 2." 
Nature Methods 9 (4). Nature Publishing Group: 357-59. 
\url{http://dx.doi.org/10.1038/nmeth.1923}.
Picard Tools - by Broad Institute. 2016.
\url{http://broadinstittue.github.io/picard/}.
}
\seealso{
\link{sample_table_galonska} \link{input_galonska} \link{TSS_galonska}
}
\keyword{datasets}
