% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DChipRep.R
\docType{package}
\name{DChIPRep}
\alias{DChIPRep}
\alias{DChIPRep-package}
\title{DChIPRep: A package for differential analysis of histone modification ChIP-Seq profiles}
\description{
The DChIPRep package provides functions to perform a differential analysis
of histone modification profiles at base-pair resolution
}
\section{DChIPRep functions}{

The \code{DChIPRep} packages provides functions for data import
\code{\link{importData}}
and  performing
position wise tests. After data import, a \code{\link{DChIPRepResults}}
object on which the function \code{\link{runTesting}}
is run to perform the tests and add the result to the object. Then, plots
can be created from this object. See the vignette for additional details:
 \code{vignette("DChIPRepVignette")}
}

