% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DChIPRepResults-class}
\alias{DChIPRepResults-class}
\alias{DChIPRepResults}
\title{DChIPRepResults object and constructor}
\usage{
DChIPRepResults(object)
}
\arguments{
\item{object}{A DESeqDataSet}
}
\value{
A DChIPRepResult object.
}
\description{
The \code{DChIPRepResults} contains a DESeqDataSet as obtained after the
initial
import.
}
\examples{
data(testData)
dcr <- DChIPRepResults(testData)

}
