% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{DESeq2Data}
\alias{DESeq2Data}
\alias{DESeq2Data,DChIPRepResults-method}
\alias{DESeq2Data<-}
\alias{DESeq2Data<-,DChIPRepResults,DESeqDataSet-method}
\title{Accessors for the 'DESeq2Data' slot of a \code{DChIPRepResults} object.}
\usage{
\S4method{DESeq2Data}{DChIPRepResults}(object)

\S4method{DESeq2Data}{DChIPRepResults,DESeqDataSet}(object) <- value
}
\arguments{
\item{object}{a \code{DChIPRepResults} object}

\item{value}{A DESeqDataSet object}
}
\value{
the DESeq2Data object contained in the DChIPRepResults object
}
\description{
The slot contains the DESeqDataSet as it is obtained after
the initial data import. The DESeqDataSet contains the
counts per position and the normalization factors as computed  using
the input counts.
}
\examples{
data(testData)
dcr <- DChIPRepResults(testData)
DESeq2Data(dcr)

}

