% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{exampleChipData}
\alias{exampleChipData}
\title{An example ChIP data.}
\format{a matrix}
\usage{
data(exampleChipData)
}
\value{
a matrix
}
\description{
An example ChIP data set that can be used with
\code{\link{importDataFromMatrices}}.
Its associated sample table can be accessed via 
data(data(exampleSampleTable).
}
\seealso{
\link{exampleSampleTable} \link{exampleInputData}
}
\keyword{datasets}
