% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{exampleInputData}
\alias{exampleInputData}
\title{An example input data.}
\format{a matrix}
\usage{
data(exampleInputData)
}
\value{
a matrix
}
\description{
An example input data set that can be used with
\code{\link{importDataFromMatrices}}.
Its associated sample table can be accessed via
data(data(exampleSampleTable).
}
\seealso{
\link{exampleSampleTable} \link{exampleChipData}
}
\keyword{datasets}
