% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{exampleSampleTable}
\alias{exampleSampleTable}
\title{An example sample table data.frame}
\format{a data.frame}
\usage{
data(exampleSampleTable)
}
\value{
a data.frame
}
\description{
An example sample table
}
\seealso{
\link{exampleChipData}   \link{exampleInputData}
}
\keyword{datasets}

