% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{FDRresults}
\alias{FDRresults}
\alias{FDRresults,DChIPRepResults-method}
\alias{FDRresults<-}
\alias{FDRresults<-,DChIPRepResults,list-method}
\title{Accessor and setter for the 'FDRresults' slot of a \code{DChIPRepResults}
object.}
\usage{
\S4method{FDRresults}{DChIPRepResults}(object)

\S4method{FDRresults}{DChIPRepResults,list}(object) <- value
}
\arguments{
\item{object}{a \code{DChIPRepResults} object}

\item{value}{A DESeqDataSet object}
}
\value{
a list containing the estimated false discovery rates
}
\description{
The slot contains the results of the FDR estimation as performed within the
function \code{\link{runTesting}}. It is the complete output of the
\code{\link[fdrtool]{fdrtool}} function.
}
\examples{
data(testData)
dcr <- DChIPRepResults(testData)
dcr <- runTesting(dcr)
str(FDRresults(dcr))
}

