% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{getMATfromDataFrame}
\alias{getMATfromDataFrame}
\title{Helper function to turn a  data.frame  into a matrix
and remove the ID column.}
\usage{
getMATfromDataFrame(df, ID = "name")
}
\arguments{
\item{df}{the input data frame with positions in the columns and the genomic
features in the rows.}

\item{ID}{the name of the ID column to be removed.}
}
\value{
a matrix with the ID column removed
}
\description{
This function takes a data.frame, with the genomic features
(e.g. transcripts or genes) in the rows and the positions upstream and
downstream of
the TSS in the columns as well as a column ID containing a genomic feature ID
and returns the data.frame with the ID column removed. The input for this
function
are tables obtained after running the Python import script.
}
\examples{
data(exampleSampleTable)
directory <- file.path(system.file("extdata", package="DChIPRep"))
df  <- lapply(file.path(directory, exampleSampleTable$Input), 
read.delim)[[1]]
mat <- getMATfromDataFrame(df)
}

