% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataImport.R
\name{importData}
\alias{importData}
\title{Import the data after running the Python script}
\usage{
importData(sampleTable, directory = "", ID = "name", ...)
}
\arguments{
\item{sampleTable}{a data.frame that has to contain the columns ChiP, Input,
sampleID,
upstream, downstream and condition. Each row of the table describes one
experimental sample. Each row of the table describes one
experimental sample. See \code{data(exampleSampleTable)} for an example table.
and the vignette for further information.}

\item{directory}{the directory relative to which the filenames are specified
given as a character.}

\item{ID}{character giving the name of the feature identifier
column in the count tables.

Defaults to \code{"name"}}

\item{...}{parameters passed to \code{\link{summarizeCountsPerPosition}}}
}
\value{
a \code{DChIPRepResults} object containg the imported data as a
\code{\link[DESeq2]{DESeqDataSet}}.
}
\description{
This function imports the data from the count table files as returned by
the accompanying Python script.
}
\examples{
data(exampleSampleTable)
directory <- file.path(system.file("extdata", package="DChIPRep"))
importedData <- importData(exampleSampleTable, directory)
}

