% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{importDataFromMatrices}
\alias{importDataFromMatrices}
\title{Import the data from ChiP and input matrices}
\usage{
importDataFromMatrices(inputData, chipData, sampleTable)
}
\arguments{
\item{inputData}{a matrix containing the counts for the input per position.}

\item{chipData}{a matrix containing the counts for the ChIP per position.}

\item{sampleTable}{a data.frame that has to contain the columns sampleID,
upstream, downstream and condition. Each row of the table describes one
experimental sample. See \code{data(exampleSampleTable)} for an example table.
and the vignette for further information.}
}
\value{
a DChIPRepResults object containing the imported data as a
\code{\link[DESeq2]{DESeqDataSet}}.
}
\description{
This function imports the data from two matrices that
contain counts summarized
per position. It computes the normalization factors from the input
(one per position)  and creates a DChIPRepResults
object.
}
\details{
The normalization factors are computed as
\code{t(t(inputData) * (covC/covI)) },  Where covC and covI contain the total
sum of the ChIP and the input samples. Zero normalization factors can arise
if the input has zero counts for certain
positions. That's why input values equal to zero are set to 1
in order to always
obtain valid normalizationFactors.
}
\examples{
data(exampleSampleTable)
data(exampleInputData)
data(exampleChipData)
imDataFromMatrices <- importDataFromMatrices(inputData = exampleInputData,
chipData = exampleChipData,
sampleTable = exampleSampleTable)
}

