% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{input_galonska}
\alias{input_galonska}
\title{Another example Input data set that can be used 
with \code{\link{importDataFromMatrices}} from rom Galonska et. al., 2015.}
\format{a matrix}
\usage{
data(input_galonska)
}
\value{
a matrix
}
\description{
The matrix containes the whole cell extract (WCE) data for H3Kme3 
from the paper
in each of the four columns, since
this is the only inpu data provided for all 4 samples. For additional
information see the documentation
of \link{chip_galonska}.
}
\seealso{
\link{chip_galonska} \link{sample_table_galonska} \link{TSS_galonska}
}
\keyword{datasets}

