% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\docType{methods}
\name{plotProfiles}
\alias{plotProfiles}
\alias{plotProfiles,DChIPRepResults-method}
\title{Produce a TSS plot of the two conditions in the data}
\usage{
\S4method{plotProfiles}{DChIPRepResults}(object, meanFunction = function(x) { 
     smhuber(x)$mu }, ...)
}
\arguments{
\item{object}{a \code{DChIPRepResults} object after \code{\link{runTesting}}}

\item{meanFunction}{a function to compute the positionwise mean per group,
defaults  to a Huber estimator of the mean.}

\item{...}{additional parametes for plotting (NOT YET IMPLEMENTED)}
}
\value{
a \code{ggplot2} object
}
\description{
This function plots the positionwise mean of the log2 of the normalized
counts of the  two conditions
after \code{\link{runTesting}} has been run on a \code{DChIPRepResults}
object.
}
\examples{
data(testData)
dcr <- DChIPRepResults(testData)
dcr <- runTesting(dcr)
plotProfiles(dcr)
}

