% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\docType{methods}
\name{plotSignificance}
\alias{plotSignificance}
\alias{plotSignificance,DChIPRepResults-method}
\title{Produce a plot that colors the positions identified as significant}
\usage{
\S4method{plotSignificance}{DChIPRepResults}(object,
  meanFunction = robust_mean, lfdrThresh = 0.2, ...)
}
\arguments{
\item{object}{a \code{DChIPRepResults} object after \code{\link{runTesting}}}

\item{meanFunction}{a function to compute the positionwise mean
per group, defaults to a Huber estimator of the mean.}

\item{lfdrThresh}{Threshold for the local FDR}

\item{...}{additional parameters for plotting (NOT YET IMPLEMENTED)}
}
\value{
a \code{ggplot2} object
}
\description{
This function plots the positionwise mean of the two conditions after
\code{\link{runTesting}}
has been run on a \code{DChIPRepResults} object. The points corresponding
to significant
positions are colored black in both  of the conditions.
The function returns the plot as a \code{ggplot2}
object that can be modified afterwards.
}
\examples{
data(testData)
dcr <- DChIPRepResults(testData)
dcr <- runTesting(dcr)
plotSignificance(dcr)
}
