% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{resultsDChIPRep}
\alias{resultsDChIPRep}
\alias{resultsDChIPRep,DChIPRepResults-method}
\alias{resultsDChIPRep<-}
\alias{resultsDChIPRep<-,DChIPRepResults,list-method}
\title{Accessors and setter for the 'results' slot of 
a \code{DChIPRepResults} object.}
\usage{
\S4method{resultsDChIPRep}{DChIPRepResults}(object)

\S4method{resultsDChIPRep}{DChIPRepResults,list}(object) <- value
}
\arguments{
\item{object}{a \code{DChIPRepResults} object}

\item{value}{A DESeqDataSet object}
}
\value{
a data.frame containing the results of the position wise tests
}
\description{
The slot contains the results of the position wise tests in a
data.frame after runing
the function \code{\link{runTesting}}. It is a modified output of the
\code{\link[DESeq2]{results}} function of the \code{DESeq2} package.
}
\examples{
data(testData)
dcr <- DChIPRepResults(testData)
dcr <- runTesting(dcr)
head(resultsDChIPRep(dcr))

}

