% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{robust_mean}
\alias{robust_mean}
\title{Use a huber type estimator to produce a robust mean}
\usage{
robust_mean(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
a robust mean of a numerical vector
}
\description{
This function uses a Huber type estimator as implemented in the function
\code{\link[smoothmest]{smhuber}} from the package
\url{https://cran.r-project.org/web/packages/smoothmest/}{smoothmest}.
This is used to summarize the profiles across replicates.
We provide a wrapper around the original function that catches the case that 
we want to produce a mean of a single value. It is used in the functions \code{\link{plotProfiles}}
and \code{\link{plotSignificance}}.
}
\examples{
data(testData)
robust_mean(counts(testData[, 1]))

x <- rcauchy(10)
robust_mean(x)
}
