% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTesting.R
\docType{methods}
\name{runTesting}
\alias{runTesting}
\alias{runTesting,DChIPRepResults-method}
\title{Run the tests on a DChIPRepResults object.}
\usage{
\S4method{runTesting}{DChIPRepResults}(object, lfcThreshold = 0.05,
  plotFDR = FALSE, ...)
}
\arguments{
\item{object}{A \link{DChIPRepResults} object.}

\item{lfcThreshold}{A non-negative threshold value, which determines the null
hypothesis. The null hypothesis is
\eqn{H_0: |log2(FC)| > } \code{lfcThreshold}}

\item{plotFDR}{If set to TRUE a plot showing the
estimated FDRs will be displayed}

\item{...}{not used currently}
}
\value{
a modified \link{DChIPRepResults} object containing the testing results
}
\description{
This function runs the testing on a \code{DChIPRepResults} object. It adds
the FDR calculations and the result table to the \code{DChIPRepResults}
object.
}
\examples{
data(testData)
dcr <- DChIPRepResults(testData)
dcr <- runTesting(dcr)
}
\seealso{
\code{\link{resultsDChIPRep}}
}

