% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{sample_table_galonska}
\alias{sample_table_galonska}
\title{Another example sample table based on data  from rom Galonska et. al., 2015.}
\format{a data.frame}
\usage{
data(sample_table_galonska)
}
\value{
a data.frame
}
\description{
This table contains the sample annotation for the H3Kme3 data from
 Galonska et. al., 2015. For additional information see the documentation
 of \link{chip_galonska}.
}
\seealso{
\link{chip_galonska}  \link{input_galonska} \link{TSS_galonska}
}
\keyword{datasets}

