% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{summarizeCountsPerPosition}
\alias{summarizeCountsPerPosition}
\title{Helper function to summarize the counts per position}
\usage{
summarizeCountsPerPosition(mat, ct = 0, trim = 0.15)
}
\arguments{
\item{mat}{the input matrix with positions in the columns and the genomic
features in the rows.}

\item{ct}{the count threshold to use.}

\item{trim}{the trimming percentage for the trimmed mean.}
}
\value{
a vector containing the summarized counts per condition
}
\description{
This function takes a matrix of counts, with the genomic features
(e.g. transcripts or genes) in the rows  and the positions upstream
and downstream of
the TSS in the columns and returns a vector with the summarized counts per
position.
}
\details{
The summary per condition is computed as a trimmed mean per position. First,
counts greater than \code{ct}  are removed and then a trimmed mean with
a trimming
percentage of \code{trim} is computed on the log scale. This mean is
then exponentiated
again and multiplied by the total number of features passing the threshold
\code{ct}
per position. If a position contains only zero counts, its mean
is returned as zero.
}
\examples{
data(exampleSampleTable)
directory <- file.path(system.file("extdata", package="DChIPRep"))
df  <- lapply(file.path(directory, exampleSampleTable$Input), 
read.delim)[[1]]
mat <- getMATfromDataFrame(df)
summaryPerPos <- summarizeCountsPerPosition(mat)
}
