% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{testData}
\alias{testData}
\title{A test DESeqDataSet}
\format{a  DESeqDataSet}
\usage{
data(testData)
}
\value{
a  DESeqDataSet
}
\description{
test data to check the functions
}
\keyword{datasets}

